move end-to-end test to separate file
